requirements.txt
flask
flask-cors
flask-socketio
numpy
Pillow
rplidar
